chore: move permission controller to sheets pacakge
