+++ packages/sheets-conditional-formatting-ui/src/controllers/cf.permission.controller.ts
@@ -17,15 +17,14 @@
 import type { ICommandInfo } from '@univerjs/core';
 import type { IAddCfCommandParams } from '../commands/commands/add-cf.command';
 import { Disposable, ICommandService, Inject, LocaleService } from '@univerjs/core';
-import { RangeProtectionPermissionEditPoint, WorkbookEditablePermission, WorksheetEditPermission, WorksheetSetCellStylePermission } from '@univerjs/sheets';
-import { SheetPermissionInterceptorBaseController } from '@univerjs/sheets-ui';
+import { RangeProtectionPermissionEditPoint, SheetPermissionCheckController, WorkbookEditablePermission, WorksheetEditPermission, WorksheetSetCellStylePermission } from '@univerjs/sheets';
 import { AddCfCommand } from '../commands/commands/add-cf.command';
 
 export class ConditionalFormattingPermissionController extends Disposable {
     constructor(
         @Inject(LocaleService) private _localeService: LocaleService,
         @ICommandService private readonly _commandService: ICommandService,
-        @Inject(SheetPermissionInterceptorBaseController) private readonly _sheetPermissionInterceptorBaseController: SheetPermissionInterceptorBaseController
+        @Inject(SheetPermissionCheckController) private readonly _SheetPermissionCheckController: SheetPermissionCheckController
 
     ) {
         super();
@@ -37,13 +36,13 @@ export class ConditionalFormattingPermissionController extends Disposable {
         this.disposeWithMe(
             this._commandService.beforeCommandExecuted((command: ICommandInfo) => {
                 if (command.id === AddCfCommand.id) {
-                    const permission = this._sheetPermissionInterceptorBaseController.permissionCheckWithRanges({
+                    const permission = this._SheetPermissionCheckController.permissionCheckWithRanges({
                         workbookTypes: [WorkbookEditablePermission],
                         rangeTypes: [RangeProtectionPermissionEditPoint],
                         worksheetTypes: [WorksheetEditPermission, WorksheetSetCellStylePermission],
                     }, (command.params as IAddCfCommandParams).rule.ranges);
                     if (!permission) {
-                        this._sheetPermissionInterceptorBaseController.haveNotPermissionHandle(this._localeService.t('permission.dialog.setStyleErr'));
+                        this._SheetPermissionCheckController.blockExecuteWithoutPermission(this._localeService.t('permission.dialog.setStyleErr'));
                     }
                 }
             })
