+++ packages/sheets-thread-comment-ui/src/controllers/sheets-thread-comment-hover.controller.ts
@@ -15,9 +15,9 @@
  */
 
 import { Disposable, Inject } from '@univerjs/core';
-import { RangeProtectionPermissionViewPoint, WorkbookCommentPermission, WorksheetViewPermission } from '@univerjs/sheets';
+import { RangeProtectionPermissionViewPoint, SheetPermissionCheckController, WorkbookCommentPermission, WorksheetViewPermission } from '@univerjs/sheets';
 import { SheetsThreadCommentModel } from '@univerjs/sheets-thread-comment';
-import { HoverManagerService, SheetPermissionInterceptorBaseController } from '@univerjs/sheets-ui';
+import { HoverManagerService } from '@univerjs/sheets-ui';
 import { debounceTime } from 'rxjs';
 import { SheetsThreadCommentPopupService } from '../services/sheets-thread-comment-popup.service';
 
@@ -26,7 +26,7 @@ export class SheetsThreadCommentHoverController extends Disposable {
         @Inject(HoverManagerService) private readonly _hoverManagerService: HoverManagerService,
         @Inject(SheetsThreadCommentPopupService) private readonly _sheetsThreadCommentPopupService: SheetsThreadCommentPopupService,
         @Inject(SheetsThreadCommentModel) private readonly _sheetsThreadCommentModel: SheetsThreadCommentModel,
-        @Inject(SheetPermissionInterceptorBaseController) private readonly _sheetPermissionInterceptorBaseController: SheetPermissionInterceptorBaseController
+        @Inject(SheetPermissionCheckController) private readonly _sheetPermissionCheckController: SheetPermissionCheckController
     ) {
         super();
         this._initHoverEvent();
@@ -43,7 +43,7 @@ export class SheetsThreadCommentHoverController extends Disposable {
                     const commentId = this._sheetsThreadCommentModel.getByLocation(unitId, subUnitId, row, col);
 
                     if (commentId) {
-                        const commentPermission = this._sheetPermissionInterceptorBaseController.permissionCheckWithRanges({
+                        const commentPermission = this._sheetPermissionCheckController.permissionCheckWithRanges({
                             workbookTypes: [WorkbookCommentPermission],
                             worksheetTypes: [WorksheetViewPermission],
                             rangeTypes: [RangeProtectionPermissionViewPoint],
