+++ packages/sheets-ui/src/plugin.ts
@@ -42,13 +42,11 @@ import { FormatPainterController } from './controllers/format-painter/format-pai
 import { HoverRenderController } from './controllers/hover-render.controller';
 import { MarkSelectionRenderController } from './controllers/mark-selection.controller';
 import { MoveRangeRenderController } from './controllers/move-range.controller';
-import { SheetPermissionInitController } from './controllers/permission/sheet-permission-init.controller';
-import { SheetPermissionInterceptorBaseController } from './controllers/permission/sheet-permission-interceptor-base.controller';
+import { SheetPermissionCheckUIController } from './controllers/permission/sheet-permission-check-ui.controller';
 import { SheetPermissionInterceptorCanvasRenderController } from './controllers/permission/sheet-permission-interceptor-canvas-render.controller';
 import { SheetPermissionInterceptorClipboardController } from './controllers/permission/sheet-permission-interceptor-clipboard.controller';
 import { SheetPermissionInterceptorFormulaRenderController } from './controllers/permission/sheet-permission-interceptor-formula-render.controller';
 import { SheetPermissionRenderController, SheetPermissionRenderManagerController, WorksheetProtectionRenderController } from './controllers/permission/sheet-permission-render.controller';
-import { SheetPermissionViewModelController } from './controllers/permission/sheet-permission-view-model.controller';
 import { SheetContextMenuRenderController } from './controllers/render-controllers/contextmenu.render-controller';
 import { EditorBridgeRenderController } from './controllers/render-controllers/editor-bridge.render-controller';
 import { FormatPainterRenderController } from './controllers/render-controllers/format-painter.render-controller';
@@ -156,14 +154,11 @@ export class UniverSheetsUIPlugin extends Plugin {
             [SheetPermissionPanelModel],
             [SheetPermissionUserManagerService],
             [SheetPermissionInterceptorClipboardController],
-            [SheetPermissionInterceptorBaseController],
-            [SheetPermissionInitController],
-            [SheetPermissionViewModelController],
+            [SheetPermissionCheckUIController],
             [SheetPermissionRenderManagerController],
         ] as Dependency[], this._config.override));
 
         touchDependencies(this._injector, [
-            [SheetPermissionViewModelController],
             [SheetPermissionPanelModel],
         ]);
     }
@@ -183,7 +178,7 @@ export class UniverSheetsUIPlugin extends Plugin {
             [SheetUIController],
             [SheetsRenderService],
             [ActiveWorksheetController],
-            [SheetPermissionInterceptorBaseController],
+            [SheetPermissionCheckUIController],
         ]);
     }
 
@@ -191,7 +186,6 @@ export class UniverSheetsUIPlugin extends Plugin {
         this._registerRenderModules();
 
         touchDependencies(this._injector, [
-            [SheetPermissionInitController],
             [SheetPermissionRenderManagerController],
             [SheetClipboardController],
             [FormulaEditorController],
