+++ packages/sheets/src/controllers/permission/sheet-permission-init.controller.ts
@@ -16,11 +16,22 @@
 
 import type { Workbook } from '@univerjs/core';
 import type { UnitAction } from '@univerjs/protocol';
-import type { IAddRangeProtectionMutationParams, IAddWorksheetProtectionParams, ISetWorksheetPermissionPointsMutationParams } from '@univerjs/sheets';
+import type { IAddRangeProtectionMutationParams } from '../../commands/mutations/add-range-protection.mutation';
+import type { IAddWorksheetProtectionParams } from '../../commands/mutations/add-worksheet-protection.mutation';
+import type { ISetWorksheetPermissionPointsMutationParams } from '../../commands/mutations/set-worksheet-permission-points.mutation';
 import { Disposable, IAuthzIoService, ICommandService, Inject, IPermissionService, IUndoRedoService, IUniverInstanceService, UniverInstanceType, UserManagerService } from '@univerjs/core';
 import { UnitObject } from '@univerjs/protocol';
-import { AddRangeProtectionMutation, AddWorksheetProtectionMutation, baseProtectionActions, defaultWorkbookPermissionPoints, defaultWorksheetPermissionPoint, getAllRangePermissionPoint, getAllWorkbookPermissionPoint, getAllWorksheetPermissionPoint, getAllWorksheetPermissionPointByPointPanel, RangeProtectionCache, RangeProtectionRuleModel, SetWorksheetPermissionPointsMutation, SheetInterceptorService, WorksheetProtectionPointModel, WorksheetProtectionRuleModel } from '@univerjs/sheets';
 import { skip } from 'rxjs';
+import { AddRangeProtectionMutation } from '../../commands/mutations/add-range-protection.mutation';
+import { AddWorksheetProtectionMutation } from '../../commands/mutations/add-worksheet-protection.mutation';
+import { SetWorksheetPermissionPointsMutation } from '../../commands/mutations/set-worksheet-permission-points.mutation';
+import { RangeProtectionRuleModel } from '../../model/range-protection-rule.model';
+import { RangeProtectionCache } from '../../model/range-protection.cache';
+import { defaultWorksheetPermissionPoint, getAllWorksheetPermissionPoint, getAllWorksheetPermissionPointByPointPanel } from '../../services/permission';
+import { baseProtectionActions, getAllRangePermissionPoint } from '../../services/permission/range-permission/util';
+import { defaultWorkbookPermissionPoints, getAllWorkbookPermissionPoint } from '../../services/permission/workbook-permission';
+import { WorksheetProtectionPointModel, WorksheetProtectionRuleModel } from '../../services/permission/worksheet-permission';
+import { SheetInterceptorService } from '../../services/sheet-interceptor/sheet-interceptor.service';
 
 export class SheetPermissionInitController extends Disposable {
     constructor(
